discord-agency-bot/requirements.txt
discord.py==2.3.2
openpyxl==3.1.2
pandas==2.2.2
apscheduler==3.10.4